refactor: move crawl handler schema to videos schema directory
