+++ src/apps/io/crawler/index.ts
@@ -1,10 +1,10 @@
 import express, { Router } from 'express';
 import { validateRequest } from 'src/utils/validator';
 import { crawlHandler } from './routes/crawl';
-import { CrawlHandlerInput, crawlHandlerSchema } from './routes/crawl/schema';
+import { crawlHandlerSchema, CrawlHandlerRequest } from 'src/schema/videos/crawl';
 
 const crawlerRouter: Router = express.Router();
 
-crawlerRouter.post('/crawl-handler', validateRequest<CrawlHandlerInput>(crawlHandlerSchema), crawlHandler);
+crawlerRouter.post('/crawl-handler', validateRequest<CrawlHandlerRequest>(crawlHandlerSchema), crawlHandler);
 
 export { crawlerRouter };
