+++ src/schema/videos/crawl/index.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
 import { hasuraEventMetadataSchema, headersSchema, transformEventMetadata, transformHeaders } from 'src/schema/hasura';
+import { taskHandlerHeaderSchema } from 'src/utils/cloud-task/schema';
 
 const CrawlRequestSchema = z.object({
   id: z.string().uuid(),
@@ -43,5 +44,24 @@ const CrawlSchema = z
     event: transformEvent(req.body.event),
   }));
 
-export { CrawlEventSchema, CrawlRequestSchema, CrawlSchema };
+const crawlHandlerSchema = z.object({
+  headers: taskHandlerHeaderSchema.passthrough(),
+  body: z.object({
+    data: z.object({
+      getSingleVideo: z.boolean(),
+      url: z.string().url(),
+      title: z.string().min(1),
+      slugPrefix: z.string().optional().default(''),
+      userId: z.string().uuid(),
+    }),
+    metadata: z.object({
+      id: hasuraEventMetadataSchema.shape.id,
+      spanId: hasuraEventMetadataSchema.shape.span_id,
+      traceId: hasuraEventMetadataSchema.shape.trace_id,
+    }),
+  }),
+});
+
+export { CrawlEventSchema, CrawlRequestSchema, CrawlSchema, crawlHandlerSchema };
 export type CrawlRequest = z.infer<typeof CrawlSchema>;
+export type CrawlHandlerRequest = z.infer<typeof crawlHandlerSchema>;
